# chello-be/src/swagger.yaml
openapi: 3.0.3
info:
  title: Chello - Task management
  description: |-
    This is an application made by using ReactJS, NodeJS, TypeScript
  # termsOfService: http://swagger.io/terms/
  # contact:
  #   email: chello@swagger.io
  # license:
  #   name: Apache 2.0
  #   url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:3001
  - url: https://chello-be-a5ll.onrender.com
  - url: https://chello-api.onrender.com
...
paths:
...
  /workspace:
    post:
      tags:
        - workspace
      summary: Create workspace
      description: This can only be done by the logged in admin.
      operationId: createWorkspace
      requestBody:
        description: Created workspace object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Workspace"
          application/xml:
            schema:
              $ref: "#/components/schemas/Workspace"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Workspace"
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
            application/xml:
              schema:
                $ref: "#/components/schemas/Workspace"
    get:
      tags:
        - workspace
      summary: Show list of workspaces
      description: This can only be done by the logged in user.
      operationId: getWorkspace
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                $items: "#/components/schemas/Workspace"
            application/xml:
              schema:
                type: array
                $items: "#/components/schemas/Workspace"
  /workspace/{workspaceId}:
    get:
      tags:
        - workspace
      summary: Get workspace and render by workspace id
      description: ""
      operationId: getWorkspaceById
      parameters:
        - name: workspaceId
          in: path
          description: "The ID that needs to be fetched"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
            application/xml:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          description: Invalid id supplied
        "404":
          description: Workspace not found
    put:
      tags:
        - workspace
      summary: Update workspace
      description: This can only be done by the logged in workspace's admin.
      operationId: updateWorkspace
      parameters:
        - name: workspaceId
          example: tF2aHCPz1NRiKTU89oqy
          in: path
          description: id that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent workspace in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Workspace"
          application/xml:
            schema:
              $ref: "#/components/schemas/Workspace"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Workspace"
      responses:
        default:
          description: successful operation
    patch:
      tags:
        - workspace
      summary: Patch workspace with body (isPublic || isFavorite)
      description: This can only be done by the logged in workspace's admin.
      operationId: patchWorkspace
      parameters:
        - name: workspaceId
          example: tF2aHCPz1NRiKTU89oqy
          in: path
          description: id that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Patch workspace with isPublic or isFavorite
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspacePatch"
          application/xml:
            schema:
              $ref: "#/components/schemas/WorkspacePatch"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/WorkspacePatch"
      responses:
        "200":
          description: Thay đổi thành công
        "400":
          description: Thay đổi thất bại
    delete:
      tags:
        - workspace
      summary: Delete workspace
      description: This can only be done by the logged in workspace's admin.
      operationId: deleteWorkspace
      parameters:
        - name: workspaceId
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted successfully
        "400":
          description: Invalid id supplied
        "404":
          description: Workspace not found
...
  /workspace/{workspaceId}/card/{cardId}:
    get:
      tags:
        - card
      summary: Get card and render by workspace and card id
      description: ""
      operationId: getCardById
      parameters:
        - name: workspaceId
          example: tF2aHCPz1NRiKTU89oqy
          in: path
          description: "The workspace ID that its card needs to be fetched"
          required: true
          schema:
            type: string
        - name: cardId
          example: E8p4qhizMaLiWI9XkXKb
          in: path
          description: "The ID that needs to be fetched"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
            application/xml:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Invalid id supplied
        "404":
          description: Card not found
    put:
      tags:
        - card
      summary: Update card
      description: This can only be done by the logged in member.
      operationId: updateCard
      parameters:
        - name: workspaceId
          example: tF2aHCPz1NRiKTU89oqy
          in: path
          description: "The workspace ID that its card needs to be updated"
          required: true
          schema:
            type: string
        - name: cardId
          example: E8p4qhizMaLiWI9XkXKb
          in: path
          description: id that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent card in the workspace
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Card"
          application/xml:
            schema:
              $ref: "#/components/schemas/Card"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Card"
      responses:
        default:
          description: successful operation
    patch:
      tags:
        - card
      summary: Patch card with field(s)
      description: This can only be done by the logged in member.
      operationId: patchCard
      parameters:
        - name: workspaceId
          example: tF2aHCPz1NRiKTU89oqy
          in: path
          description: "The workspace ID that its card field(s) needs to be updated"
          required: true
          schema:
            type: string
        - name: cardId
          example: E8p4qhizMaLiWI9XkXKb
          in: path
          description: card id that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update card with field(s)
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardPatch"
          application/xml:
            schema:
              $ref: "#/components/schemas/CardPatch"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/CardPatch"
      responses:
        "200":
          description: Thay đổi thành công
        "400":
          description: Thay đổi thất bại
    delete:
      tags:
        - card
      summary: Delete card
      description: This can only be done by the logged in workspace's admin.
      operationId: deleteWorkspace
      parameters:
        - name: workspaceId
          in: path
          description: "The workspace ID that its card needs to be deleted"
          required: true
          schema:
            type: string
        - name: cardId
          in: path
          description: The card id that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted successfully
        "400":
          description: Invalid id supplied
        "404":
          description: Workspace not found
...
components:
  schemas:
    User:
      required:
        - userId
        - mail
        - password
      type: object
      properties:
        userId:
          type: string
          example: "user10"
        username:
          type: string
          example: theUser
        mail:
          type: string
          example: john@email.com
        password:
          type: string
          example: "12345"
        avatar:
          type: string
        # phone:
        #   type: string
        #   example: "12345"
        # userStatus:
        #   type: integer
        #   description: User Status
        #   format: int32
        #   example: 1
      xml:
        name: user
    Workspace:
      required:
        # - workspaceId
        - workspaceName
      type: object
      properties:
        # id:
        #   type: string
        #   example: "ws11"
        workspaceName:
          type: string
          example: "Chello Project"
        isPublic:
          type: boolean
        isFavorite:
          type: boolean
        workspaceImage:
          type: string
          example: "https://digitalsynopsis.com/wp-content/uploads/2017/07/beautiful-color-ui-gradients-backgrounds-relay.png"
        member:
          type: array
          $items: "#/components/schemas/User"
        cards:
          type: array
          $items: "#/components/schemas/Card"
      xml: workspace
    Card:
      required:
        # - cardId
        - cardName
      type: object
      properties:
        # cardId:
        #   type: string
        #   example: "card10"
        cardName:
          type: string
          example: "Tasks"
        isActived:
          type: boolean
        items:
          type: array
          $items: "#/components/schemas/Item"
      xml: card
    Item:
      required:
        # - itemId
        - itemName
      type: object
      properties:
        # itemId:
        #   type: string
        #   example: "item4"
        itemName:
          type: string
          example: "ERD design"
        description:
          type: string
          example: "Thiết kế bằng draw.io"
        label:
          type: string
          example: "Label example"
        background:
          type: "string"
          example: "red"
        deadline:
          type: string
          format: date
        member:
          type: array
          $items: "#/components/schemas/User"
        todoList:
          type: array
          $items: "#/components/schemas/TodoList"
      xml: item
    TodoList:
      required:
        # - todoListId
        - description
      type: object
      properties:
        # todoListId:
        #   type: string
        #   example: "todo01"
        description:
          type: string
          example: "Bước 1 là bay"
        isDone:
          type: boolean
      xml: todoList
...
    WorkspacePatch:
      type: object
      properties:
        isPublic:
          type: boolean
        isFavorite:
          type: boolean
      xml:
        name: "workspace"
    CardPatch:
      type: object
      properties:
        isActived:
          type: boolean
      xml:
        name: "card"
  # requestBodies:
  #   Pet:
  #     description: Pet object that needs to be added to the store
  #     content:
  #       application/json:
  #         schema:
  #           $ref: "#/components/schemas/Pet"
  #       application/xml:
  #         schema:
  #           $ref: "#/components/schemas/Pet"
  #   UserArray:
  #     description: List of user object
  #     content:
  #       application/json:
  #         schema:
  #           type: array
  #           items:
  #             $ref: "#/components/schemas/User"
  # securitySchemes:
  #   petstore_auth:
  #     type: oauth2
  #     flows:
  #       implicit:
  #         authorizationUrl: https://petstore3.swagger.io/oauth/authorize
  #         scopes:
  #           write:pets: modify pets in your account
  #           read:pets: read your pets
  #   api_key:
  #     type: apiKey
  #     name: api_key
  #     in: header